created a new file getTime where we have all function relate with time and coundown days
