+++ src/client/index.js
@@ -2,3 +2,4 @@ import './sass/main.scss'
 
 import './js/app'
 import './js/handleNav'
+import './js/getTime'
